@@ -1,7 +1,7 @@
 {
   "name": "mongoose",
   "description": "Mongoose MongoDB ODM",
-  "version": "8.16.4",
+  "version": "8.19.1",
   "author": "Guillermo Rauch <guillermo@learnboost.com>",
   "keywords": [
     "mongodb",
@@ -22,15 +22,16 @@
   "dependencies": {
     "bson": "^6.10.4",
     "kareem": "2.6.3",
-    "mongodb": "~6.17.0",
+    "mongodb": "~6.20.0",
     "mpath": "0.9.0",
     "mquery": "5.0.0",
     "ms": "2.1.3",
     "sift": "17.1.3"
   },
   "devDependencies": {
-    "@babel/core": "7.27.7",
-    "@babel/preset-env": "7.27.2",
+    "@ark/attest": "0.48.2",
+    "@babel/core": "7.28.4",
+    "@babel/preset-env": "7.28.3",
     "@mongodb-js/mongodb-downloader": "^0.4.2",
     "@typescript-eslint/eslint-plugin": "^8.19.1",
     "@typescript-eslint/parser": "^8.19.1",
@@ -42,7 +43,7 @@
     "babel-loader": "8.2.5",
     "broken-link-checker": "^0.7.8",
     "buffer": "^5.6.0",
-    "cheerio": "1.1.0",
+    "cheerio": "1.1.2",
     "crypto-browserify": "3.12.1",
     "dox": "1.0.0",
     "eslint": "8.57.1",
@@ -56,9 +57,9 @@
     "markdownlint-cli2": "^0.18.1",
     "marked": "15.0.12",
     "mkdirp": "^3.0.1",
-    "mocha": "11.7.1",
+    "mocha": "11.7.4",
     "moment": "2.30.1",
-    "mongodb-memory-server": "10.1.4",
+    "mongodb-memory-server": "10.2.1",
     "mongodb-runner": "^5.8.2",
     "ncp": "^2.0.0",
     "nyc": "15.1.0",
@@ -66,10 +67,10 @@
     "q": "1.5.1",
     "sinon": "21.0.0",
     "stream-browserify": "3.0.0",
-    "tsd": "0.32.0",
-    "typescript": "5.8.3",
+    "tsd": "0.33.0",
+    "typescript": "5.9.3",
     "uuid": "11.1.0",
-    "webpack": "5.99.9"
+    "webpack": "5.102.0"
   },
   "directories": {
     "lib": "./lib/mongoose"
@@ -111,7 +112,8 @@
     "test-encryption": "mocha --exit ./test/encryption/*.test.js",
     "tdd": "mocha ./test/*.test.js --inspect --watch --recursive --watch-files ./**/*.{js,ts}",
     "test-coverage": "nyc --reporter=html --reporter=text npm test",
-    "ts-benchmark": "cd ./benchmarks/typescript/simple && npm install && npm run benchmark | node ../../../scripts/tsc-diagnostics-check"
+    "ts-benchmark": "cd ./benchmarks/typescript/simple && npm install && npm run benchmark | node ../../../scripts/tsc-diagnostics-check",
+    "attest-benchmark": "node ./benchmarks/typescript/infer.bench.mts"
   },
   "main": "./index.js",
   "types": "./types/index.d.ts",
